refactor: consolidate localNoon into conf.LocalNoon
Address review feedback (CodeRabbit): the localNoon helper was duplicated
across four packages (conf, classifier, processor, species). Export it once
as conf.LocalNoon and call it from every site. conf is a leaf package that
all four already import (or can import without a cycle), and it already owns
the range-filter date logic (ShouldUpdateRangeFilterToday).

Move the localNoon unit test to internal/conf and drop the per-package copies.
No behavior change.
