+++ internal/classifier/birdnet.go
@@ -1474,7 +1474,7 @@ func (bn *BirdNET) GetSpeciesOccurrenceAtTime(species string, detectionTime time
 	// local calendar day (matching getCachedSpeciesScores, which keys on the
 	// local DateOnly of detectionTime) rather than UTC-truncating, so the
 	// fallback computes the same geomodel week as the cache path.
-	day := localNoon(detectionTime)
+	day := conf.LocalNoon(detectionTime)
 	speciesScores, err := bn.GetProbableSpecies(day, 0.0)
 	if err != nil {
 		bn.Debug("Error getting probable species for occurrence: %v", err)
