+++ internal/conf/range_filter.go
@@ -95,13 +95,14 @@ func (s *Settings) IsSpeciesIncluded(result string) bool {
 	return false
 }
 
-// localNoon returns 12:00:00 on the calendar day of t, evaluated in t's own
-// time zone. time.Time.Truncate operates on absolute (UTC) time, so truncating
-// to a 24h boundary rounds to UTC midnight and rolls the "today" marker over at
-// UTC midnight rather than local midnight. Anchoring on the local calendar day
-// makes the daily range-filter update roll over at local midnight; noon (rather
-// than midnight) also sidesteps zones whose DST transition happens at 00:00.
-func localNoon(t time.Time) time.Time {
+// LocalNoon returns 12:00:00 on the calendar day of t, evaluated in t's own time
+// zone. Range-filter date logic uses it to anchor "today" on the local calendar
+// day: time.Time.Truncate operates on absolute (UTC) time, so truncating to a 24h
+// boundary rounds to UTC midnight and, near the local day boundary on hosts with a
+// non-zero UTC offset, yields the wrong calendar day for the geomodel week lookup
+// (and rolls the daily-update marker over at UTC rather than local midnight). Using
+// noon rather than midnight also sidesteps zones whose DST transition is at 00:00.
+func LocalNoon(t time.Time) time.Time {
 	return time.Date(t.Year(), t.Month(), t.Day(), 12, 0, 0, 0, t.Location())
 }
 
@@ -121,7 +122,7 @@ func ShouldUpdateRangeFilterToday() bool {
 		return false
 	}
 
-	today := localNoon(time.Now())
+	today := LocalNoon(time.Now())
 	// Compare local-day to local-day. UpdateIncludedSpecies stamps LastUpdated with
 	// the wall-clock time.Now() of the rebuild, which is earlier than today's noon
 	// anchor for any rebuild that finishes in the local morning. Comparing the raw
@@ -129,7 +130,7 @@ func ShouldUpdateRangeFilterToday() bool {
 	// triggering a full geomodel rebuild each time. Bucketing LastUpdated to its own
 	// local noon keeps the "only the first caller on a given local day gets true"
 	// guarantee (issue #1357) regardless of what time the last rebuild ran.
-	if !localNoon(current.BirdNET.RangeFilter.LastUpdated).Before(today) {
+	if !LocalNoon(current.BirdNET.RangeFilter.LastUpdated).Before(today) {
 		return false
 	}
